renamed sponsorships component to adBar and fixed the sponsors service to make sure it is responsible for placeholders
