+++ frontend/src/app/components/adBar/adBar.component.scss
