+++ frontend/src/app/components/adBar/adBar.component.spec.ts
