--- frontend/src/app/components/sponsorships/sponsorships.component.html
@@ -1,10 +1,0 @@
-<div class="logos">
-  <div class="logos-slide">
-    <img *ngFor="let sponsor of sponsors" [src]="(sponsor.imageUrl)">
-  </div>
-
-  <div class="logos-slide">
-    <img *ngFor="let sponsor of sponsors" [src]="(sponsor.imageUrl)">
-  </div>
-</div>
-<div class="overlay"></div>
